requirements.txt
asgiref==3.8.1
Django==5.1.6
django-ckeditor-5==0.2.17
django-extensions==4.0
pillow==11.1.0
sqlparse==0.5.3
whitenoise==6.6.0
gunicorn
boto3
django-storages

